네이티브 쿼리 qlrm 을 사용해서 구독리스트 가져오기 백엔드 완료

+++ src/main/java/com/example/photogram/web/dto/subscribe/SubscribeDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,6 +16,9 @@ public class SubscribeDto { //구독 모달창
     private int id; //누군가의 정보 구독할 취소 할 id값
     private String username;
     private String profileImageUrl; //모달창 구독한 유저 사진
-    private Integer subscribeState; //구독상태
-    private Integer equalUserState; //로그인한 사용자의 동일인 여부
+
+    //MySQL에서 COUNT 같은 연산 결과를 반환 할떄는 BigInteger를 활용해야 한다
+    private BigInteger subscribeState; //구독상태
+    private BigInteger equalUserState; //로그인한 사용자의 동일인 여부
+
 }
